requirements.txt
Flask==0.12.2
requests-toolbelt
google-cloud-storage
GoogleAppEngineCloudStorageClient